+++ src/alpaka/plugin-SiPixelClusterizer/clustering_alpaka.cc
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "../AlpakaCore/alpakaConfig.h"
+#include "AlpakaCore/alpakaConfig.h"
 
 int main() {
   std::cout << "Exit success" << std::endl;
